// src/payload-types.ts
export interface Page {
  id: number;
  title: string;
  description: string;
  slug: string;
  layout: (
    | {
        callUsButton: {
          label: string;
          url: string;
        };
        tableDiscount: {
          tableHeader?: string | null;
          table: {
            root: {
              type: string;
              children: {
                type: string;
                version: number;
                [k: string]: unknown;
              }[];
              direction: ('ltr' | 'rtl') | null;
              format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
              indent: number;
              version: number;
            };
            [k: string]: unknown;
          };
        };
        tableRegular: {
          table: {
            root: {
              type: string;
              children: {
                type: string;
                version: number;
                [k: string]: unknown;
              }[];
              direction: ('ltr' | 'rtl') | null;
              format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
              indent: number;
              version: number;
            };
            [k: string]: unknown;
          };
        };
        rules?: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'home';
      }
    | {
        phone: {
          label: string;
          url: string;
        };
        workSchedule?: {
          title?: string | null;
          workTime?: {
            monday?: string | null;
            tuesday?: string | null;
            wednesday?: string | null;
            thursday?: string | null;
            friday?: string | null;
            saturday?: string | null;
            sunday?: string | null;
          };
        };
        address: string;
        routeButton: {
          label: string;
          url: string;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'contacts';
      }
    | {
        rules: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'rules';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}